Added a permissions check before any queries are made. The microservice also returns the debug value now

+++ DuggaSys/microservices/duggaedService/deleteDugga_ms.php
@@ -2,6 +2,7 @@
 
 include_once "../../../Shared/sessions.php";
 include_once "../../../Shared/basic.php";
+include_once "../sharedMicroservices/getUid_ms.php";
 
 date_default_timezone_set("Europe/Stockholm");
 
@@ -11,20 +12,27 @@ session_start();
 
 $opt=getOP('opt');
 $qid=getOP('qid');
+$cid=getOP('cid');
+$debug = "NONE!";
+$userid=getUid();
 
-if(strcmp($opt,"DELDU")===0){
-	$query = $pdo->prepare("DELETE FROM userAnswer WHERE quiz=:qid;");
-	$query->bindParam(':qid', $qid);
-	if(!$query->execute()) {
-		$debug = "Useranswer could not be deleted.";
-	}
+if(checklogin() && (hasAccess($userid, $cid, 'w') || isSuperUser($userid) || hasAccess($userid, $cid, 'st'))){
+	if(strcmp($opt,"DELDU")===0){
+		$query = $pdo->prepare("DELETE FROM userAnswer WHERE quiz=:qid;");
+		$query->bindParam(':qid', $qid);
+		if(!$query->execute()) {
+			$debug = "Useranswer could not be deleted.";
+		}
 
-	$query = $pdo->prepare("DELETE FROM quiz WHERE id=:qid;");
-	$query->bindParam(':qid', $qid);
+		$query = $pdo->prepare("DELETE FROM quiz WHERE id=:qid;");
+		$query->bindParam(':qid', $qid);
 
-	if(!$query->execute()) {
-		$debug = "Quiz could not be deleted.";
+		if(!$query->execute()) {
+			$debug = "Quiz could not be deleted.";
+		}
 	}
 }
+
+echo json_encode($debug);
 
 ?>
